feat: Introduce repository dialog components and refactor registry table actions
- Added new dialogs for syncing, deleting, and selecting commits for repositories.
- Created a centralized RepositoryActions component for handling repository actions in the table.
- Introduced an ActiveDialog enum to manage dialog states effectively.
- Refactored the RegistryRepositoriesTable component to utilize the new dialog components and improve state management.
- Enhanced user experience by providing clear actions for repository management directly in the table.

This update streamlines the repository management process and improves the overall UI consistency.
