+++ frontend/src/components/registry/registry-common.tsx
@@ -1,0 +1,5 @@
+export enum ActiveDialog {
+  RepositorySync,
+  RepositoryDelete,
+  RepositoryCommit,
+}
