+++ frontend/src/components/registry/registry-repos-table.tsx
@@ -4,15 +4,10 @@ import { DotsHorizontalIcon } from "@radix-ui/react-icons"
 import {
   AlertTriangleIcon,
   ChevronDownIcon,
-  CopyIcon,
   CornerDownRightIcon,
-  GitBranchIcon,
   LoaderCircleIcon,
-  RefreshCcw,
-  Trash2Icon,
-  TrashIcon,
 } from "lucide-react"
-import { useCallback, useState } from "react"
+import { useState } from "react"
 import type {
   RegistryRepositoryErrorDetail,
   RegistryRepositoryReadMinimal,
@@ -22,17 +17,11 @@ import {
   DataTableColumnHeader,
   type DataTableToolbarProps,
 } from "@/components/data-table"
-import { CommitSelector } from "@/components/registry/commit-selector"
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-} from "@/components/ui/alert-dialog"
+import { CommitSelectorDialog } from "@/components/registry/dialogs/repository-commit-dialog"
+import { DeleteRepositoryDialog } from "@/components/registry/dialogs/repository-delete-dialog"
+import { SyncRepositoryDialog } from "@/components/registry/dialogs/repository-sync-dialog"
+import { ActiveDialog } from "@/components/registry/registry-common"
+import { RepositoryActions } from "@/components/registry/table-actions"
 import { Badge } from "@/components/ui/badge"
 import { Button } from "@/components/ui/button"
 import {
@@ -41,29 +30,14 @@ import {
   CollapsibleTrigger,
 } from "@/components/ui/collapsible"
 import {
-  Dialog,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog"
-import {
   DropdownMenu,
   DropdownMenuContent,
-  DropdownMenuItem,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu"
 import { TooltipProvider } from "@/components/ui/tooltip"
-import { toast } from "@/components/ui/use-toast"
 import { useAuth } from "@/hooks/use-auth"
 import { getRelativeTime } from "@/lib/event-history"
-import { useRegistryRepositories, useRepositoryCommits } from "@/lib/hooks"
-
-enum AlertAction {
-  SYNC,
-  DELETE,
-  SYNC_EXECUTOR,
-  COMMIT_SELECT,
-}
+import { useRegistryRepositories } from "@/lib/hooks"
 
 export function RegistryRepositoriesTable() {
   const { user } = useAuth()
@@ -71,150 +45,21 @@ export function RegistryRepositoriesTable() {
     repos: registryRepos,
     reposIsLoading: registryReposIsLoading,
     reposError: registryReposError,
-    syncRepo,
-    syncRepoIsPending,
-    deleteRepo,
     syncRepoError,
+    syncRepoIsPending,
   } = useRegistryRepositories()
+
+  // Dialog state management
+  const [activeDialog, setActiveDialog] = useState<ActiveDialog | null>(null)
   const [selectedRepo, setSelectedRepo] =
     useState<RegistryRepositoryReadMinimal | null>(null)
-  const [alertAction, setAlertAction] = useState<AlertAction | null>(null)
-  const [alertOpen, setAlertOpen] = useState(false)
-  const [commitDialogOpen, setCommitDialogOpen] = useState(false)
-  const [selectedCommitSha, setSelectedCommitSha] = useState<string | null>(
-    null
-  )
 
-  // Fetch commits for the selected repository
-  const { commits, commitsIsLoading, commitsError } = useRepositoryCommits(
-    selectedRepo?.id || null,
-    { enabled: commitDialogOpen }
-  )
-  const getAlertContent = useCallback(() => {
-    switch (alertAction) {
-      case AlertAction.SYNC:
-        return {
-          title: "Sync repository",
-          description: (
-            <span className="flex flex-col space-y-2">
-              <span>
-                You are about to pull the latest version of the repository{" "}
-              </span>
-              <b className="font-mono tracking-tighter">
-                {selectedRepo?.origin}
-              </b>
-              {selectedRepo?.commit_sha && (
-                <span className="text-sm text-muted-foreground">
-                  <span>Current SHA: </span>
-                  <Badge className="font-mono text-xs" variant="secondary">
-                    {selectedRepo.commit_sha}
-                  </Badge>
-                </span>
-              )}
-              {selectedRepo?.last_synced_at && (
-                <span className="text-sm text-muted-foreground">
-                  <span>Last synced: </span>
-                  <span>
-                    {new Date(selectedRepo.last_synced_at).toLocaleString()}
-                  </span>
-                </span>
-              )}
-              <span>
-                Are you sure you want to proceed? This will reload all existing
-                actions with the latest versions from the remote repository.
-              </span>
-            </span>
-          ),
-          actions: [
-            {
-              label: (
-                <div className="flex items-center space-x-2">
-                  <RefreshCcw className="size-4" />
-                  <span>Sync</span>
-                </div>
-              ),
-              action: async () => {
-                if (!selectedRepo) {
-                  console.error("No repository selected")
-                  return
-                }
-                console.log("Reloading repository", selectedRepo.origin)
-                try {
-                  await syncRepo({ repositoryId: selectedRepo.id })
-                  toast({
-                    title: "Successfully synced repository",
-                    description: (
-                      <span className="flex flex-col space-y-2">
-                        <span>
-                          Successfully reloaded actions from{" "}
-                          <b className="inline-block">{selectedRepo.origin}</b>
-                        </span>
-                      </span>
-                    ),
-                  })
-                } catch (error) {
-                  console.error("Error reloading repository", error)
-                } finally {
-                  setSelectedRepo(null)
-                }
-              },
-            },
-          ],
-        }
-      case AlertAction.DELETE:
-        return {
-          title: "Delete repository",
-          description: (
-            <span className="flex flex-col space-y-2">
-              <span>You are about to delete the repository </span>
-              <b className="font-mono tracking-tighter">
-                {selectedRepo?.origin}
-              </b>
-              <span>
-                Are you sure you want to proceed? This action cannot be undone.
-              </span>
-              <span className="italic">
-                You cannot delete the base Tracecat actions or the custom
-                template repositories. If you delete your remote repository, you
-                will need to restart the instance to restore it.
-              </span>
-            </span>
-          ),
-          actions: [
-            {
-              label: (
-                <div className="flex items-center space-x-2">
-                  <Trash2Icon className="size-4" />
-                  <span>Delete</span>
-                </div>
-              ),
-              action: async () => {
-                if (!selectedRepo) {
-                  console.error("No repository selected")
-                  return
-                }
-                console.log(
-                  "Deleting repository",
-                  selectedRepo.origin,
-                  selectedRepo.id
-                )
-                try {
-                  await deleteRepo({ repositoryId: selectedRepo.id })
-                } catch (error) {
-                  console.error("Error deleting repository", error)
-                } finally {
-                  setSelectedRepo(null)
-                }
-              },
-            },
-          ],
-        }
-      default:
-        return null
+  const onOpenChange = (open: boolean) => {
+    if (!open) {
+      setActiveDialog(null)
+      setSelectedRepo(null)
     }
-  }, [alertAction, selectedRepo, syncRepo, deleteRepo])
-
-  const alertContent = getAlertContent()
+  }
 
   const errorDetail =
     syncRepoError?.status === 422
@@ -280,383 +125,168 @@ export function RegistryRepositoriesTable() {
           </Collapsible>
         </div>
       )}
-      <AlertDialog open={alertOpen}>
-        <DataTable
-          isLoading={registryReposIsLoading}
-          error={registryReposError ?? undefined}
-          data={registryRepos}
-          emptyMessage="No actions found."
-          errorMessage="Error loading workflows."
-          columns={[
-            {
-              accessorKey: "origin",
-              header: ({ column }) => (
-                <DataTableColumnHeader
-                  className="text-xs"
-                  column={column}
-                  title="Origin"
-                />
-              ),
-              cell: ({ row }) => {
-                return (
-                  <div className="text-xs text-foreground/80">
-                    {row.getValue<RegistryRepositoryReadMinimal["origin"]>(
-                      "origin"
-                    ) || "-"}
-                  </div>
-                )
-              },
-              enableSorting: true,
-              enableHiding: false,
-            },
-            {
-              accessorKey: "commit_sha",
-              header: ({ column }) => (
-                <DataTableColumnHeader
-                  className="text-xs"
-                  column={column}
-                  title="Commit SHA"
-                />
-              ),
-              cell: ({ row }) => {
-                const sha =
-                  row.getValue<RegistryRepositoryReadMinimal["commit_sha"]>(
-                    "commit_sha"
-                  )
-                if (!sha)
-                  return <div className="text-xs text-foreground/80">-</div>
-                return (
-                  <Badge
-                    className="font-mono text-xs font-normal"
-                    variant="secondary"
-                  >
-                    {sha.substring(0, 7)}
-                  </Badge>
-                )
-              },
-              enableHiding: false,
-            },
-            {
-              accessorKey: "last_synced_at",
-              header: ({ column }) => (
-                <DataTableColumnHeader
-                  className="text-xs"
-                  column={column}
-                  title="Last synced"
-                />
-              ),
-              cell: ({ row }) => {
-                const lastSyncedAt =
-                  row.getValue<RegistryRepositoryReadMinimal["last_synced_at"]>(
-                    "last_synced_at"
-                  )
-                if (!lastSyncedAt) {
-                  return <div className="text-xs text-foreground/80">-</div>
-                }
-                const date = new Date(lastSyncedAt)
-                const ago = getRelativeTime(date)
-                return (
-                  <div className="space-x-2 text-xs">
-                    <span>{date.toLocaleString()}</span>
-                    <span className="text-muted-foreground">({ago})</span>
-                  </div>
-                )
-              },
-              enableSorting: true,
-              enableHiding: false,
-            },
-            {
-              id: "actions",
-              enableHiding: false,
-              cell: ({ row }) => {
-                const commitSha = row.original.commit_sha
-                const isSelected = row.original.id === selectedRepo?.id
-                const Icon = () => {
-                  if (isSelected && syncRepoIsPending) {
-                    return <LoaderCircleIcon className="size-4 animate-spin" />
-                  }
-                  const errorDetail = syncRepoError?.body
-                    .detail as RegistryRepositoryErrorDetail
-                  if (errorDetail?.id === row.original.id) {
-                    return (
-                      <AlertTriangleIcon className="size-4 fill-rose-600 text-white" />
-                    )
-                  }
-                  return <DotsHorizontalIcon className="size-4" />
-                }
-                return (
-                  <DropdownMenu>
-                    <DropdownMenuTrigger asChild>
-                      <Button
-                        variant="ghost"
-                        className="size-8 p-0"
-                        onClick={(e) => e.stopPropagation()} // Prevent row click
-                      >
-                        <span className="sr-only">Open menu</span>
-                        <Icon />
-                      </Button>
-                    </DropdownMenuTrigger>
-                    <DropdownMenuContent>
-                      <DropdownMenuItem
-                        className="flex items-center text-xs"
-                        onClick={(e) => {
-                          e.stopPropagation() // Prevent row click
-                          navigator.clipboard.writeText(row.original.origin)
-                          toast({
-                            title: "Repository origin copied",
-                            description: (
-                              <span className="flex flex-col space-y-2">
-                                <span className="inline-block">
-                                  {row.original.origin}
-                                </span>
-                              </span>
-                            ),
-                          })
-                        }}
-                      >
-                        <CopyIcon className="mr-2 size-4" />
-                        <span>Copy repository origin</span>
-                      </DropdownMenuItem>
-                      {commitSha !== null && (
-                        <DropdownMenuItem
-                          className="flex items-center text-xs"
-                          onClick={(e) => {
-                            e.stopPropagation() // Prevent row click
-                            navigator.clipboard.writeText(commitSha)
-                            toast({
-                              title: "Commit SHA copied",
-                              description: (
-                                <span className="flex flex-col space-y-2">
-                                  <span className="inline-block">
-                                    {commitSha}
-                                  </span>
-                                </span>
-                              ),
-                            })
-                          }}
-                        >
-                          <CopyIcon className="mr-2 size-4" />
-                          <span>Copy commit SHA</span>
-                        </DropdownMenuItem>
-                      )}
-                      {user?.isOrgAdmin() && (
-                        <>
-                          <DropdownMenuItem
-                            className="flex items-center text-xs"
-                            onClick={(e) => {
-                              e.stopPropagation() // Prevent row click
-                              setSelectedRepo(row.original)
-                              setAlertAction(AlertAction.SYNC)
-                              setAlertOpen(true)
-                            }}
-                          >
-                            <RefreshCcw className="mr-2 size-4" />
-                            <span>Sync from remote</span>
-                          </DropdownMenuItem>
-                          {row.original.origin.startsWith("git+ssh://") && (
-                            <DropdownMenuItem
-                              className="flex items-center text-xs"
-                              onClick={(e) => {
-                                e.stopPropagation() // Prevent row click
-                                setSelectedRepo(row.original)
-                                setSelectedCommitSha(row.original.commit_sha)
-                                setCommitDialogOpen(true)
-                              }}
-                            >
-                              <GitBranchIcon className="mr-2 size-4" />
-                              <span>Change commit</span>
-                            </DropdownMenuItem>
-                          )}
-                          <DropdownMenuItem
-                            className="flex items-center text-xs text-rose-600"
-                            onClick={(e) => {
-                              e.stopPropagation() // Prevent row click
-                              setSelectedRepo(row.original)
-                              setAlertAction(AlertAction.DELETE)
-                              setAlertOpen(true)
-                            }}
-                          >
-                            <TrashIcon className="mr-2 size-4" />
-                            <span>Delete repository</span>
-                          </DropdownMenuItem>
-                        </>
-                      )}
-                    </DropdownMenuContent>
-                  </DropdownMenu>
-                )
-              },
-            },
-          ]}
-          toolbarProps={defaultToolbarProps}
-        />
-        <AlertDialogContent>
-          <AlertDialogHeader>
-            <AlertDialogTitle>{alertContent?.title}</AlertDialogTitle>
-            <AlertDialogDescription>
-              {alertContent?.description}
-            </AlertDialogDescription>
-          </AlertDialogHeader>
-          <AlertDialogFooter>
-            <AlertDialogCancel onClick={() => setAlertOpen(false)}>
-              Cancel
-            </AlertDialogCancel>
-
-            {alertContent?.actions.map((action, index) => (
-              <AlertDialogAction
-                key={index}
-                onClick={async () => {
-                  setAlertOpen(false)
-                  await action.action()
-                }}
-                disabled={syncRepoIsPending}
-              >
-                {action.label}
-              </AlertDialogAction>
-            ))}
-          </AlertDialogFooter>
-        </AlertDialogContent>
-      </AlertDialog>
-
-      {/* Commit Selection Dialog */}
-      <Dialog open={commitDialogOpen} onOpenChange={setCommitDialogOpen}>
-        <DialogContent className="max-w-2xl">
-          <DialogHeader>
-            <DialogTitle>Select commit to sync</DialogTitle>
-          </DialogHeader>
-          <div className="space-y-4">
-            {selectedRepo && (
-              <div>
-                <p className="text-sm text-muted-foreground mb-4">
-                  Select a commit to sync repository:{" "}
-                  <span className="font-mono">{selectedRepo.origin}</span>
-                </p>
-                <div className="space-y-4">
-                  <CommitSelector
-                    commits={commits}
-                    currentCommitSha={selectedCommitSha}
-                    isLoading={commitsIsLoading}
-                    error={commitsError}
-                    onSelectCommit={(commitSha) => {
-                      setSelectedCommitSha(commitSha)
-                    }}
-                  />
-
-                  {/* Selected Commit Details */}
-                  {selectedCommitSha &&
-                    commits &&
-                    (() => {
-                      const selectedCommit = commits.find(
-                        (commit) => commit.sha === selectedCommitSha
-                      )
-                      if (!selectedCommit) return null
-
-                      const commitDate = new Date(selectedCommit.date)
-                      const relativeTime = getRelativeTime(commitDate)
-                      const isHead = commits[0]?.sha === selectedCommitSha
-
-                      return (
-                        <div className="rounded-lg border bg-muted/50 p-4 space-y-3">
-                          <div className="flex items-center justify-between">
-                            <h4 className="font-medium text-sm">
-                              Selected commit details
-                            </h4>
-                            <div className="flex items-center space-x-2">
-                              <Badge
-                                variant="secondary"
-                                className="font-mono text-xs"
-                              >
-                                {selectedCommit.sha.substring(0, 7)}
-                              </Badge>
-                              {isHead && (
-                                <Badge variant="default" className="text-xs">
-                                  HEAD
-                                </Badge>
-                              )}
-                            </div>
-                          </div>
-                          <div className="space-y-2">
-                            <div>
-                              <p className="text-sm font-medium text-foreground">
-                                {selectedCommit.message}
-                              </p>
-                            </div>
-                            <div className="flex items-center justify-between text-xs text-muted-foreground">
-                              <span>
-                                by {selectedCommit.author} (
-                                {selectedCommit.author_email})
-                              </span>
-                              <div className="flex items-center space-x-2">
-                                <span>{commitDate.toLocaleDateString()}</span>
-                                <span>•</span>
-                                <span>{relativeTime}</span>
-                              </div>
-                            </div>
-                            <div className="pt-1">
-                              <span className="text-xs text-muted-foreground font-mono">
-                                Full SHA: {selectedCommit.sha}
-                              </span>
-                            </div>
-                          </div>
-                        </div>
-                      )
-                    })()}
+      <DataTable
+        isLoading={registryReposIsLoading}
+        error={registryReposError ?? undefined}
+        data={registryRepos}
+        emptyMessage="No actions found."
+        errorMessage="Error loading workflows."
+        columns={[
+          {
+            accessorKey: "origin",
+            header: ({ column }) => (
+              <DataTableColumnHeader
+                className="text-xs"
+                column={column}
+                title="Origin"
+              />
+            ),
+            cell: ({ row }) => {
+              return (
+                <div className="text-xs text-foreground/80">
+                  {row.getValue<RegistryRepositoryReadMinimal["origin"]>(
+                    "origin"
+                  ) || "-"}
                 </div>
-              </div>
-            )}
-            <div className="flex justify-end space-x-2 pt-4">
-              <Button
-                variant="outline"
-                onClick={() => {
-                  setCommitDialogOpen(false)
-                  setSelectedCommitSha(null)
-                }}
-              >
-                Cancel
-              </Button>
-              <Button
-                onClick={async () => {
-                  if (!selectedRepo || !selectedCommitSha) return
+              )
+            },
+            enableSorting: true,
+            enableHiding: false,
+          },
+          {
+            accessorKey: "commit_sha",
+            header: ({ column }) => (
+              <DataTableColumnHeader
+                className="text-xs"
+                column={column}
+                title="Commit SHA"
+              />
+            ),
+            cell: ({ row }) => {
+              const sha =
+                row.getValue<RegistryRepositoryReadMinimal["commit_sha"]>(
+                  "commit_sha"
+                )
+              if (!sha)
+                return <div className="text-xs text-foreground/80">-</div>
+              return (
+                <Badge
+                  className="font-mono text-xs font-normal"
+                  variant="secondary"
+                >
+                  {sha.substring(0, 7)}
+                </Badge>
+              )
+            },
+            enableHiding: false,
+          },
+          {
+            accessorKey: "last_synced_at",
+            header: ({ column }) => (
+              <DataTableColumnHeader
+                className="text-xs"
+                column={column}
+                title="Last synced"
+              />
+            ),
+            cell: ({ row }) => {
+              const lastSyncedAt =
+                row.getValue<RegistryRepositoryReadMinimal["last_synced_at"]>(
+                  "last_synced_at"
+                )
+              if (!lastSyncedAt) {
+                return <div className="text-xs text-foreground/80">-</div>
+              }
+              const date = new Date(lastSyncedAt)
+              const ago = getRelativeTime(date)
+              return (
+                <div className="space-x-2 text-xs">
+                  <span>{date.toLocaleString()}</span>
+                  <span className="text-muted-foreground">({ago})</span>
+                </div>
+              )
+            },
+            enableSorting: true,
+            enableHiding: false,
+          },
+          {
+            id: "actions",
+            enableHiding: false,
+            cell: ({ row }) => {
+              const isRepoSelected = selectedRepo?.id === row.original.id
+              const Icon = () => {
+                if (isRepoSelected && syncRepoIsPending) {
+                  return <LoaderCircleIcon className="size-4 animate-spin" />
+                }
+                const errorDetail = syncRepoError?.body
+                  .detail as RegistryRepositoryErrorDetail
+                if (errorDetail?.id === row.original.id) {
+                  return (
+                    <AlertTriangleIcon className="size-4 fill-rose-600 text-white" />
+                  )
+                }
+                return <DotsHorizontalIcon className="size-4" />
+              }
+              return (
+                <DropdownMenu>
+                  <DropdownMenuTrigger asChild>
+                    <Button
+                      variant="ghost"
+                      className="size-8 p-0"
+                      onClick={(e) => e.stopPropagation()} // Prevent row click
+                    >
+                      <span className="sr-only">Open menu</span>
+                      <Icon />
+                    </Button>
+                  </DropdownMenuTrigger>
+                  <DropdownMenuContent>
+                    {user?.isOrgAdmin() && (
+                      <RepositoryActions
+                        repository={row.original}
+                        onSync={() => {
+                          setSelectedRepo(row.original)
+                          setActiveDialog(ActiveDialog.RepositorySync)
+                        }}
+                        onDelete={() => {
+                          setSelectedRepo(row.original)
+                          setActiveDialog(ActiveDialog.RepositoryDelete)
+                        }}
+                        onChangeCommit={() => {
+                          setSelectedRepo(row.original)
+                          setActiveDialog(ActiveDialog.RepositoryCommit)
+                        }}
+                      />
+                    )}
+                  </DropdownMenuContent>
+                </DropdownMenu>
+              )
+            },
+          },
+        ]}
+        toolbarProps={defaultToolbarProps}
+      />
 
-                  try {
-                    await syncRepo({
-                      repositoryId: selectedRepo.id,
-                      requestBody: {
-                        target_commit_sha: selectedCommitSha,
-                      },
-                    })
-                    setCommitDialogOpen(false)
-                    setSelectedCommitSha(null)
-                    toast({
-                      title: "Successfully synced repository",
-                      description: (
-                        <span className="flex flex-col space-y-2">
-                          <span>
-                            Successfully synced{" "}
-                            <b className="inline-block">
-                              {selectedRepo.origin}
-                            </b>
-                          </span>
-                          <span className="text-xs">
-                            to commit {selectedCommitSha.substring(0, 7)}
-                          </span>
-                        </span>
-                      ),
-                    })
-                  } catch (error) {
-                    console.error(
-                      "Error syncing repository to specific commit",
-                      error
-                    )
-                  }
-                }}
-                disabled={syncRepoIsPending || !selectedCommitSha}
-              >
-                {syncRepoIsPending ? "Syncing..." : "Sync"}
-              </Button>
-            </div>
-          </div>
-        </DialogContent>
-      </Dialog>
+      {/* Dialog Components */}
+      <SyncRepositoryDialog
+        open={activeDialog === ActiveDialog.RepositorySync}
+        onOpenChange={onOpenChange}
+        selectedRepo={selectedRepo}
+        setSelectedRepo={setSelectedRepo}
+      />
+
+      <DeleteRepositoryDialog
+        open={activeDialog === ActiveDialog.RepositoryDelete}
+        onOpenChange={onOpenChange}
+        selectedRepo={selectedRepo}
+        setSelectedRepo={setSelectedRepo}
+      />
+
+      <CommitSelectorDialog
+        open={activeDialog === ActiveDialog.RepositoryCommit}
+        onOpenChange={onOpenChange}
+        selectedRepo={selectedRepo}
+        initialCommitSha={selectedRepo?.commit_sha}
+      />
     </TooltipProvider>
   )
 }
